Laravel 7 upgrade, new CS, new CI scripts, more test coverage more coverage + removal of dead code use QueueableAction in all actions remove all registration related logic remove unused class add coverage for project model updated/recently updated logic upgrade packages + remove sanitize from marked as it's depricated prettier autofixes prettier is handled via eslint Revert "prettier is handled via eslint"
This reverts commit ef8b890f9130e9bcac00e7280aa2e03b03aad3a6.
no prettier command for js, this is done through eslint

+++ database/migrations/2018_08_24_211441_create_permission_tables.php
@@ -1,30 +1,32 @@
 <?php
 
+declare(strict_types=1);
+
 use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Support\Facades\App;
 use Illuminate\Support\Facades\Schema;
 
-/**
- * Creates the permission tables.
- */
 class CreatePermissionTables extends Migration
 {
-    /**
-     * Run the migrations.
-     */
-    public function up()
+    public function up(): void
     {
         $tableNames = config('permission.table_names');
         $columnNames = config('permission.column_names');
 
-        Schema::create($tableNames['permissions'], function (Blueprint $table) {
+        if (empty($tableNames)) {
+            throw new \Exception('Error: config/permission.php not loaded. Run [php artisan config:clear] '
+                . 'and try again.');
+        }
+
+        Schema::create($tableNames['permissions'], function (Blueprint $table): void {
             $table->increments('id');
             $table->string('name');
             $table->string('guard_name');
             $table->timestamps();
         });
 
-        Schema::create($tableNames['roles'], function (Blueprint $table) {
+        Schema::create($tableNames['roles'], function (Blueprint $table): void {
             $table->increments('id');
             $table->string('name');
             $table->string('guard_name');
@@ -33,13 +35,16 @@ class CreatePermissionTables extends Migration
 
         Schema::create(
             $tableNames['model_has_permissions'],
-            function (Blueprint $table) use ($tableNames, $columnNames) {
+            function (Blueprint $table) use ($tableNames, $columnNames): void {
                 $table->unsignedInteger('permission_id');
 
                 $table->string('model_type');
                 $table->unsignedBigInteger($columnNames['model_morph_key']);
                 $table->index(
-                    [$columnNames['model_morph_key'], 'model_type'],
+                    [
+                        $columnNames['model_morph_key'],
+                        'model_type',
+                    ],
                     'model_has_permissions_model_id_model_type_index'
                 );
 
@@ -57,7 +62,7 @@ class CreatePermissionTables extends Migration
 
         Schema::create(
             $tableNames['model_has_roles'],
-            function (Blueprint $table) use ($tableNames, $columnNames) {
+            function (Blueprint $table) use ($tableNames, $columnNames): void {
                 $table->unsignedInteger('role_id');
 
                 $table->string('model_type');
@@ -79,40 +84,38 @@ class CreatePermissionTables extends Migration
             }
         );
 
-        Schema::create(
-            $tableNames['role_has_permissions'],
-            function (Blueprint $table) use ($tableNames) {
-                $table->unsignedInteger('permission_id');
-                $table->unsignedInteger('role_id');
+        Schema::create($tableNames['role_has_permissions'], function (Blueprint $table) use ($tableNames): void {
+            $table->unsignedInteger('permission_id');
+            $table->unsignedInteger('role_id');
 
-                $table->foreign('permission_id')
-                    ->references('id')
-                    ->on($tableNames['permissions'])
-                    ->onDelete('cascade');
+            $table->foreign('permission_id')
+                ->references('id')
+                ->on($tableNames['permissions'])
+                ->onDelete('cascade');
 
-                $table->foreign('role_id')
-                    ->references('id')
-                    ->on($tableNames['roles'])
-                    ->onDelete('cascade');
+            $table->foreign('role_id')
+                ->references('id')
+                ->on($tableNames['roles'])
+                ->onDelete('cascade');
 
-                $table->primary(
-                    ['permission_id', 'role_id'],
-                    'role_has_permissions_permission_id_role_id_primary'
-                );
-            }
-        );
+            $table->primary(['permission_id', 'role_id'], 'role_has_permissions_permission_id_role_id_primary');
+        });
 
-        app('cache')
+        App::get('cache')
             ->store(config('permission.cache.store') != 'default' ? config('permission.cache.store') : null)
             ->forget(config('permission.cache.key'));
     }
 
-    /**
-     * Reverse the migrations.
-     */
-    public function down()
+    public function down(): void
     {
         $tableNames = config('permission.table_names');
+
+        if (empty($tableNames)) {
+            throw new \Exception(
+                'Error: config/permission.php not found and defaults could not be merged.'
+                . ' Please publish the package configuration before proceeding, or drop the tables manually.'
+            );
+        }
 
         Schema::drop($tableNames['role_has_permissions']);
         Schema::drop($tableNames['model_has_roles']);
